Bug fixes, and stuff

+++ Cryptaxation/Models/K4Transaction.cs
@@ -10,9 +10,9 @@ namespace Cryptaxation
     {
         public decimal Amount { get; set; }
         public string Currency { get; set; }
-        public decimal SalesPrice { get; set; }
-        public decimal TaxBasis { get; set; }
-        public decimal Gain { get; set; }
-        public decimal Loss { get; set; }
+        public int SalesPrice { get; set; }
+        public int TaxBasis { get; set; }
+        public int Gain { get; set; }
+        public int Loss { get; set; }
     }
 }
